authorized feature

+++ project/blog/app/Http/Controllers/ArticleController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -33,13 +36,18 @@ class ArticleController extends Controller
 
     public function delete(string $id) {
         $article = Article::find($id);
-        $article->delete();
-
-        return redirect("/articles")->with("info", "An article is deleted.");
+        if(Gate::allows("delete-article", $article)) {
+            $article->delete();
+            return redirect("/articles")->with("info", "An article is deleted.");
+        } else {
+            return back()->with("info", "Unauthorized to delete");
+        }
     }
 
     public function add() {
-        return view("articles.add");
+        return view("articles.add", [
+            "categories" => Category::all(),
+        ]);
     }
 
     public function create() {
@@ -58,6 +66,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = Auth::id();
         $article->save();
 
         return redirect("/articles");
